Separated outputs, fixed tests for not linux platform
